requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==4.2.7
django-ckeditor==6.7.1
django-cors-headers==4.3.0
django-extensions==3.2.3
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.1
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
persiantools==3.0.1
Pillow==10.1.0
psycopg==3.1.12
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.10.3
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
